+++ src/app/locales/repairs/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./repairs.factory";
+export * from "./repairs.util";
+export * from "./registry";